Кеш содержимого архивов
  Карта записей архива читается один раз на файл: страницы просмотра
  больше не открывают zip на каждый запрос, а перебор несуществующих
  номеров отбивается по карте — этим заняты боты. Картинки из архива
  отдаются с ETag, повторный запрос закрывает браузер.

  Начисление баллов за сообщение и статью переживает отсутствующую
  настройку модуля — раньше падало с ошибкой.

+++ Blog/Observers/ArticleObserver.php
@@ -35,8 +35,9 @@ class ArticleObserver
 
         if ($article->wasChanged('active')) {
             $user = $article->user;
-            $pointAmount = setting('blog_point');
-            $moneyAmount = setting('blog_money');
+            // Настройки может не быть — модуль ставят раньше, чем задают его настройки
+            $pointAmount = (int) setting('blog_point');
+            $moneyAmount = (int) setting('blog_money');
 
             if ($article->active) {
                 $user->increment('point', $pointAmount);
